+++ src/main/java/com/kata/books/services/AutorService.java
@@ -3,6 +3,8 @@ package com.kata.books.services;
 import com.kata.books.domain.Autor;
 import com.kata.books.domain.dtos.AutorDTO;
 import com.kata.books.repositories.AutorRepository;
+import com.kata.books.repositories.LivroRepository;
+import com.kata.books.services.exceptions.DataIntegrityViolationException;
 import com.kata.books.services.exceptions.ObjectnotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,9 @@ public class AutorService {
 
 	@Autowired
 	private AutorRepository repository;
+
+	@Autowired
+	private LivroRepository livroRepository;
 
 	public Autor findById(Integer id) {
 		Optional<Autor> obj = repository.findById(id);
@@ -41,6 +46,11 @@ public class AutorService {
 
 	public void delete(Integer id) {
 		Autor obj = findById(id);
+
+		boolean existeLivro = livroRepository.existsByAssunto_Id(id);
+		if (existeLivro) {
+			throw new DataIntegrityViolationException("Não é possível excluir: existem livros vinculados a este assunto! Id: " + id);
+		}
 		repository.delete(obj);
 	}
 }
